# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: GreenHerb API
  description: |
    API REST para a gestão de planos de cultivo da estufa GreenHerb.
    Projeto académico — Sprint 2.
  version: 0.1.0
  contact:
    name: Equipa GreenHerb
servers:
  - url: http://localhost:5000
    description: Servidor de desenvolvimento local
...
paths:
...
  /api/utilizadores:
    get:
      tags: [Utilizadores]
      summary: Listar utilizadores (apenas Administrador)
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Lista de utilizadores
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Utilizador'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
    post:
      tags: [Utilizadores]
      summary: Criar utilizador (apenas Administrador)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Utilizador criado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Utilizador'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/utilizadores/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          example: 65a1b2c3d4e5f60718293a4b
    put:
      tags: [Utilizadores]
      summary: Atualizar utilizador (apenas Administrador, exceto a própria conta)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                nome:
                  type: string
                email:
                  type: string
                  format: email
                perfil:
                  type: string
                  enum: [Técnico, Responsável, Administrador]
                ativo:
                  type: boolean
      responses:
        '200':
          description: Utilizador atualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Utilizador'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Tentativa de alterar a própria conta
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /api/tarefas:
    get:
      tags: [Tarefas]
      summary: Listar tarefas
      security:
        - bearerAuth: []
      parameters:
        - name: loteId
          in: query
          schema: { type: string }
        - name: estado
          in: query
          schema:
            type: string
            enum: [pendente, executada]
      responses:
        '200':
          description: Lista de tarefas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Tarefa'
    post:
      tags: [Tarefas]
      summary: Criar tarefa
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Tarefa'
      responses:
        '201':
          description: Tarefa criada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tarefa'
  /api/tarefas/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string }
    get:
      tags: [Tarefas]
      summary: Detalhe da tarefa
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Tarefa
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tarefa'
    put:
      tags: [Tarefas]
      summary: Editar tarefa
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Tarefa'
      responses:
        '200':
          description: Tarefa atualizada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tarefa'
    delete:
      tags: [Tarefas]
      summary: Eliminar tarefa (Administrador)
      security:
        - bearerAuth: []
      responses:
        '204': { description: Eliminada }
  /api/tarefas/{id}/executar:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string }
    put:
      tags: [Tarefas]
      summary: Marcar tarefa como executada
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Tarefa atualizada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tarefa'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    BadRequest:
      description: Pedido inválido
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Autenticação requerida ou token inválido
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Sem permissões para a operação
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Recurso não encontrado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    InternalError:
      description: Erro interno do servidor
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      properties:
        erro:
          type: string
          example: Mensagem descritiva do erro.
    Utilizador:
      type: object
      properties:
        _id:
          type: string
          example: 65a1b2c3d4e5f60718293a4b
        nome:
          type: string
          example: Ana Administradora
        email:
          type: string
          format: email
          example: admin@greenherb.local
        perfil:
          type: string
          enum: [Técnico, Responsável, Administrador]
          example: Administrador
        ativo:
          type: boolean
          example: true
        dataCriacao:
          type: string
          format: date-time
    RegisterRequest:
      type: object
      required: [nome, email, password]
      properties:
        nome:
          type: string
          example: Tânia Técnica
        email:
          type: string
          format: email
          example: tecnico@greenherb.local
        password:
          type: string
          format: password
          minLength: 6
          example: tecnico123
        perfil:
          type: string
          enum: [Técnico, Responsável, Administrador]
          default: Técnico
...
    Tarefa:
      type: object
      required: [tipo, loteId, dataPrevista]
      properties:
        _id: { type: string }
        tipo:
          type: string
          enum: [rega, fertilização, colheita, monitorização]
        loteId: { type: string }
        estado:
          type: string
          enum: [pendente, executada]
        dataPrevista: { type: string, format: date-time }
        dataExecucao: { type: string, format: date-time }
        notas: { type: string }
        criadoPor:
          oneOf:
            - type: string
            - $ref: '#/components/schemas/Utilizador'
        executadoPor:
          oneOf:
            - type: string
            - $ref: '#/components/schemas/Utilizador'
        dataCriacao: { type: string, format: date-time }
...